FIX: 즐겨찾기 페이지 로그인 유도 수정
